backend/app/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2 
pymupdf==1.25.5
python-dotenv==1.1.0 
google-generativeai==0.8.5